Cria task run_subflow
