--- pipelines/teste__flow_com_subflow/.dockerignore
@@ -1,12 +1,0 @@
-.direnv/
-.env
-.envrc
-.git/
-.github/
-.mergify.yml
-.venv
-Dockerfile
-README.md
-docs/
-flake.lock
-flake.nix
