--- pipelines/teste__flow_com_subflow/CHANGELOG.md
@@ -1,3 +1,0 @@
-# Changelog - teste__flow_com_subflow
-
-## [1.0.0] - yyyy-mm-dd
