--- pipelines/teste__flow_com_subflow/pyproject.toml
@@ -1,4 +1,0 @@
-[project]
-name = "teste__flow_com_subflow"
-version = "0.1.0"
-description = "Pipeline flow_com_subflow da secretaria rj-smtr"
