--- pipelines/teste__subflow/CHANGELOG.md
@@ -1,3 +1,0 @@
-# Changelog - teste__subflow
-
-## [1.0.0] - yyyy-mm-dd
