--- pipelines/teste__subflow/prefect.yaml
@@ -1,51 +1,0 @@
-name: teste__subflow
-prefect-version: 3.4.8
-
-build:
-  - prefect.deployments.steps.run_shell_script:
-      id: get-commit-hash
-      script: git rev-parse --short HEAD
-      stream_output: false
-  - prefect_docker.deployments.steps.build_docker_image:
-      id: build-image
-      requires: prefect-docker>=0.6.5
-      image_name: ghcr.io/rj-smtr/pipelines_v3/deployments
-      tag: "teste__subflow-{{ get-commit-hash.stdout }}"
-      dockerfile: pipelines/teste__subflow/Dockerfile
-
-push:
-  - prefect_docker.deployments.steps.push_docker_image:
-      requires: prefect-docker>=0.6.5
-      image_name: "{{ build-image.image_name }}"
-      tag: "{{ build-image.tag }}"
-
-pull:
-  - prefect.deployments.steps.set_working_directory:
-      directory: /opt/prefect/pipelines_v3
-
-deployments:
-  - name: rj-teste--subflow--staging
-    version: "{{ build-image.tag }}"
-    entrypoint: pipelines/teste__subflow/flow.py:teste__subflow
-    work_pool:
-      name: smtr-pool
-      work_queue_name: default
-      job_variables:
-        image: "{{ build-image.image_name }}:{{ build-image.tag }}"
-        command: uv run --package teste__subflow -- prefect flow-run execute
-        secretName: prefect-jobs-secrets-staging
-        image_pull_policy: Always
-  - name: rj-teste--subflow--prod
-    version: "{{ get-commit-hash.stdout }}"
-    entrypoint: pipelines/teste__subflow/flow.py:teste__subflow
-    schedules:
-      - cron: "0 0 * * *"
-        timezone: "America/Sao_Paulo"
-    work_pool:
-      name: smtr-pool
-      work_queue_name: default
-      job_variables:
-        image: "{{ build-image.image_name }}:{{ build-image.tag }}"
-        command: uv run --package teste__subflow -- prefect flow-run execute
-        secretName: prefect-jobs-secrets
-        image_pull_policy: Always
