--- pipelines/teste__subflow/pyproject.toml
@@ -1,4 +1,0 @@
-[project]
-name = "teste__subflow"
-version = "0.1.0"
-description = "Pipeline subflow da secretaria rj-smtr"
